// packages/datadog-api-client-v1/models/MonitorOptionsAggregation.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Type of aggregation performed in the monitor query.
*/
export declare class MonitorOptionsAggregation {
    /**
    * Group to break down the monitor on.
    */
    'groupBy'?: string;
    /**
    * Metric name used in the monitor.
    */
    'metric'?: string;
    /**
    * Metric type used in the monitor.
    */
    'type'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): MonitorOptionsAggregation;
    static serialize(data: MonitorOptionsAggregation): {
        [key: string]: any;
    };
    constructor();
}