// packages/datadog-api-client-v1/models/MonitorThresholds.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* List of the different monitor threshold available.
*/
export declare class MonitorThresholds {
    /**
    * The monitor `CRITICAL` threshold.
    */
    'critical'?: number;
    /**
    * The monitor `CRITICAL` recovery threshold.
    */
    'criticalRecovery'?: number;
    /**
    * The monitor `OK` threshold.
    */
    'ok'?: number;
    /**
    * The monitor UNKNOWN threshold.
    */
    'unknown'?: number;
    /**
    * The monitor `WARNING` threshold.
    */
    'warning'?: number;
    /**
    * The monitor `WARNING` recovery threshold.
    */
    'warningRecovery'?: number;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): MonitorThresholds;
    static serialize(data: MonitorThresholds): {
        [key: string]: any;
    };
    constructor();
}