// packages/datadog-api-client-v1/models/MonitorThresholdWindowOptions.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Alerting time window options.
*/
export declare class MonitorThresholdWindowOptions {
    /**
    * Describes how long an anomalous metric must be normal before the alert recovers.
    */
    'recoveryWindow'?: string;
    /**
    * Describes how long a metric must be anomalous before an alert triggers.
    */
    'triggerWindow'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): MonitorThresholdWindowOptions;
    static serialize(data: MonitorThresholdWindowOptions): {
        [key: string]: any;
    };
    constructor();
}